pyproject.toml
[tool.poetry]
name = "helm-inspect"
version = "0.0.0"
description = "Drift detection for helm resources"
readme = "README.md"
authors = ["Qreater <qreater@gmail.com>"]
license = "Apache-2.0"
keywords = ["helm", "kubernetes", "inspect", "diff"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pyyaml = "^6.0"
art = "^5.2"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"

[tool.poetry-dynamic-versioning]
enable = true

[tool.poetry.scripts]
helm-inspect = "helm_inspect.main:main"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
